Add render tests for swipe element
